If there are remote changes and no local changes, try to fast forward.
This will try to merge the remote branch into the local branch if there
are no local changes. This will perform a 'fast forward only' merge, so
there will be no merge conflicts and no merge commit.

If there are any local changes, then it is still up to the user to merge
themselves, which is the only sane way to handle that situation.