Add default remote support.
You can now set the key 'default' to true in a remote. Doing this means
that you don't have to set the remote name for each repository.

If you only set one remote, then we will use that as the default, even
if you do not explicitly set it as the default remote.