Refactor find_config_file in RepoConfigParser.
This function now makes sense.

It will now look for a .polygamy.json file, or a .polygamy directory
containing a .polygamy.json file. The function will check the parent
repositories until it reaches the root path. If the configuration files
are still not found, then a value error will be raised.